+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The format is based on [Keep a Changelog](https://kSSeepachangelog.com/en/1.1.0/).
 
-## Unreleased
+## [0.1.1] - 2026-04-02
 
 ### Added